refactor: 1.created styles file. 2.moved styled-components to styles file.
