+++ src/styles/ProcessBox/ProcessBox.tsx
